==> Aprendizaje Supervisado/Informe de Laboratorio/Scripts/1_Comprensión_del_problema.ipynb <==
# En primer lugar leemos la información del documento Excel

# Libreria util para trabajar con dataframes
import pandas as pd

dataFrame = pd.read_csv('../Material/Database.csv', sep=';')

# Extraer información de interés
import numpy as np

classes, frec = np.unique(dataFrame.Class, return_counts=True)
# Si se diera el caso de que nuestro conjunto de datos no estuviera balanceado
# tendríamos que aplicar alguna técnica de balanceo

# Seleccionar una muestra aleatoria de nuestra base de datos
import random
import cv2
import matplotlib.pyplot as plt

# Escogemos un numero aleatorio entre 0 y la longitud del dataframe
num = random.randint(0, len(dataFrame))

name_img = dataFrame.ID[num]

img = cv2.imread('../Material/Images/' + name_img)

plt.imshow(img, cmap='gray')
plt.show()

# Para entender un poquito más el problema, podemos cargar tambien las mascaras de la retina y de la RNFL
# y hacer un plot sobre la imagen para ver las regiones de interes

# Leemos las máscaras y ploteamos los contornos sobre la imagen original

# Le ponemos un 0 para trabajar con una imagenes en escala de grises. De otra manera la deblovería en 
# formato BGR
rnfl_mask = cv2.imread('../Material/RNFL_masks/' + name_img, 0)
retina_mask = cv2.imread('../Material/Retina_masks/' + name_img, 0)

# Encontramos los contornos
cont_rnfl, _ = cv2.findContours(rnfl_mask, cv2.RETR_TREE, cv2.CHAIN_APPROX_SIMPLE)
cont_retina, _ = cv2.findContours(retina_mask, cv2.RETR_TREE, cv2.CHAIN_APPROX_SIMPLE)

# Dibujamos sobre la imagen original los contornos de las mascaras 
image = cv2.drawContours(img.copy(), cont_rnfl, -1, (255, 0, 0), 3)
image = cv2.drawContours(image.copy(), cont_retina, -1, (0, 255, 0), 3)

plt.imshow(image, cmap='gray')
plt.show()
