==> Aprendizaje Supervisado/Informe de Laboratorio/Scripts/2_Particion_datos_externa.ipynb <==
# Primero leemos el dcumento csv 

import pandas as pd
import numpy as np

# Leer la información del documento Excel original

dataFrame = pd.read_csv('../Material/Database.csv', sep=';')

# Partición externa 80% - 20% de los datos
from sklearn.model_selection import train_test_split, KFold

# Partición externa tipo hold out 80% - 20% 
# train, test = train_test_split(dataFrame, test_size=0.2, shuffle=True, random_state=42)

# Partición externa tipo validación cruzada
kf = KFold(n_splits=5, shuffle=True, random_state=42)
bolsas = kf.split(dataFrame)

for k, (train_fold, test_fold) in enumerate(bolsas):
    train = dataFrame.iloc[train_fold]
    test = dataFrame.iloc[test_fold]

# Podemos hacer mas robusta nuestra partición desordenando los dataframes

train = train.sample(frac=1)
test = test.sample(frac=1)

# Por ultimo, guardamos nuestras particiones para no tener que estar llamado constantemente
# el dataFrame original

# Guardar los modelos

import os

if not os.path.exists('partitions'):
    os.mkdir('../partitions')

train.to_csv('../partitions/train.csv', sep=';')
test.to_csv('../partitions/test.csv', sep=';')
