==> Aprendizaje Supervisado/Informe de Laboratorio/Scripts/6_Evaluación_resultados.ipynb <==
# Cargar los datos del test

import numpy as np

test = np.load('../final_features/test.npy')

# El conjunto de test ya ha sido previamente estandarizado
# y ya hemos la selección de los atributos
X_test = test[:, :-1]
y_test = test[:, -1]

# cargar los modelos ya entrenados

import pickle
import sklearn import metrics
from tabulate import tabulate

with open('../models/LOGR.pickle', 'rb') as fr:
    LOGR = pickle.load(fr)

with open('../models/MLP.pickle', 'rb') as fr:
    MLP = pickle.load(fr)
    
# Extraer las predicciones
y_pred_LOGR = LOGR.predict(X_test)
y_pred_MLP = MLP.predict(X_test)

y_pred = [y_pred_LOGR, y_pred_MLP]

# Evaluar diversas métricas de clasificación
headers = ['', 'LOGR', 'MLP']
P, S, FS, ACC, AUC = [['Precission'], ['Sensibilidad'], ['F1-Score'], ['Accuracy'], ['AUC']]

for i in range(0, 2):
    P.append(np.round(metrics.preccision_score(y_test, y_pred[i]), 4))
    S.append(np.round(metrics.recall_score(y_test, y_pred[i]), 4))
    FS.append(np.round(metrics.f1_score(y_test, y_pred[i]), 4))
    ACC.append(np.round(metrics.accuracy_score(y_test, y_pred[i]), 4))
    AUC.append(np.round(metrics.roc_auc_score(y_test, y_pred[i]), 4))
    
my_data = [tuple(P), tuple(S), tuple(FS), tuple(ACC), tuple(AUC)]

# Confusion matrix
print(metrics.confussion_matrix(y_test, y_pred[1]))
